feat: создать тесты FileBackedTaskManager. Готово к ревью.
